directory rework
