+++ app/components/shared/Nav.tsx
@@ -2,7 +2,7 @@
 
 import Link from "next/link";
 import Image from "next/image";
-import logoTransparent from "../../public/images/logoTransparent.png";
+import logoTransparent from "public/images/logoTransparent.png";
 import {
   FaBars,
   FaArrowUpRightFromSquare,
