+++ app/past/2023-10-12/components/Talk.tsx
