+++ app/team/components/Developer.tsx
